% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colAnys,dgCMatrix-method}
\alias{colAnys,dgCMatrix-method}
\alias{rowAnys,dgCMatrix-method}
\title{Check if any elements in a row (column) of a matrix-like object is equal to a value}
\usage{
\S4method{colAnys}{dgCMatrix}(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

\S4method{rowAnys}{dgCMatrix}(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{value}{the value to search for.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{logical}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Check if any elements in a row (column) of a matrix-like object is equal to a value
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowAnys}}
/ \code{matrixStats::\link[matrixStats]{colAnys}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowAnys(mat)
  colAnys(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowAnys}()} and
\code{matrixStats::\link[matrixStats]{colAnys}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For checks if any element is equal to a value, see \code{\link[MatrixGenerics]{rowAnys}()}.
\item \link{any}
}
}
