% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colCollapse,dgCMatrix-method}
\alias{colCollapse,dgCMatrix-method}
\alias{rowCollapse,dgCMatrix-method}
\title{Extract one cell from each row (column) of a matrix-like object}
\usage{
\S4method{colCollapse}{dgCMatrix}(x, idxs = idxs, cols = NULL, ...)

\S4method{rowCollapse}{dgCMatrix}(x, idxs, rows = NULL, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{idxs}{An index \code{\link[base]{vector}} with the position to extract.
It is recycled to match the number of rows (column)}

\item{cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{...}{Additional arguments passed to specific methods.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Extract one cell from each row (column) of a matrix-like object
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowCollapse}}
/ \code{matrixStats::\link[matrixStats]{colCollapse}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowCollapse(mat, idxs = 2)
  rowCollapse(mat, idxs = c(1,1,2,3,2))
  
  colCollapse (mat, idxs = 4)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowCollapse}()} and
\code{matrixStats::\link[matrixStats]{colCollapse}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
}
}
