% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{colCummins,dgCMatrix-method}
\alias{colCummins,dgCMatrix-method}
\title{Calculates the cumulative minima for each row (column) of a matrix-like object}
\usage{
\S4method{colCummins}{dgCMatrix}(x, rows = NULL, cols = NULL, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{matrix}} with the
same dimensions as \code{x}.
}
\description{
Calculates the cumulative minima for each row (column) of a matrix-like object
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowCummins}}
/ \code{matrixStats::\link[matrixStats]{colCummins}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowCummins(mat)
  colCummins(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowCummins}()} and
\code{matrixStats::\link[matrixStats]{colCummins}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For single minimum estimates, see \code{\link[MatrixGenerics]{rowMins}()}.
\item \link{cummin}
}
}
