% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colDiffs,dgCMatrix-method}
\alias{colDiffs,dgCMatrix-method}
\alias{rowDiffs,dgCMatrix-method}
\title{Calculates the difference between each element of a row (column) of a
matrix-like object}
\usage{
\S4method{colDiffs}{dgCMatrix}(x, rows = NULL, cols = NULL, lag = 1L, differences = 1L, ...)

\S4method{rowDiffs}{dgCMatrix}(x, rows = NULL, cols = NULL, lag = 1L, differences = 1L, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{lag}{An integer specifying the lag.}

\item{differences}{An integer specifying the order of difference.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{matrix}}
with one column (row) less than x: \eqn{Nx(K-1)} or \eqn{(N-1)xK}.
}
\description{
Calculates the difference between each element of a row (column) of a
matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowDiffs}}
/ \code{matrixStats::\link[matrixStats]{colDiffs}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowDiffs(mat)
  colDiffs(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowDiffs}()} and
\code{matrixStats::\link[matrixStats]{colDiffs}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item \code{base::\link[base]{diff}()}.
}
}
