% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\docType{methods}
\name{colDiffs}
\alias{colDiffs}
\alias{colDiffs,dgCMatrix-method}
\alias{rowDiffs}
\alias{rowDiffs,dgCMatrix-method}
\title{Calculates difference for each row (column) in a matrix}
\usage{
colDiffs(x, rows = NULL, cols = NULL, lag = 1L, differences = 1L,
  ...)

\S4method{colDiffs}{dgCMatrix}(x, rows = NULL, cols = NULL, lag = 1L,
  differences = 1L, ...)

rowDiffs(x, rows = NULL, cols = NULL, lag = 1L, differences = 1L,
  ...)

\S4method{rowDiffs}{dgCMatrix}(x, rows = NULL, cols = NULL, lag = 1L,
  differences = 1L, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{lag}{An \code{\link[base]{integer}} specifying the lag.}

\item{differences}{An \code{\link[base]{integer}} specifying the order of
difference.}

\item{...}{Not used.}
}
\value{
Returns a \code{\link[base]{numeric}} Nx(K-1) or (N-1)xK
\code{\link[base]{matrix}}.
}
\description{
Calculates difference for each row (column) in a matrix.
}
\examples{
x <- matrix(1:27, ncol = 3)

d1 <- rowDiffs(x)
print(d1)

d2 <- t(colDiffs(t(x)))
stopifnot(all.equal(d2, d1))
}
\seealso{
See also \code{\link{diff2}}().
}
