% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\docType{methods}
\name{colIQRs}
\alias{colIQRs}
\alias{colIQRs,dgCMatrix-method}
\alias{rowIQRs}
\alias{rowIQRs,dgCMatrix-method}
\title{Estimates of the interquartile range for each row (column) in a matrix}
\usage{
colIQRs(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colIQRs}{dgCMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)

rowIQRs(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowIQRs}{dgCMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, missing values are dropped
first, otherwise not.}

\item{...}{Additional arguments passed to \code{\link{rowQuantiles}}()
(\code{colQuantiles()}).}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Estimates of the interquartile range for each row (column) in a matrix.
}
\section{Missing values}{

 Contrary to \code{\link[stats]{IQR}}, which gives
an error if there are missing values and \code{na.rm = FALSE}, \code{iqr()}
and its corresponding row and column-specific functions return
\code{\link[base]{NA}}_real_.

}

\examples{
set.seed(1)

x <- matrix(rnorm(50 * 40), nrow = 50, ncol = 40)
str(x)

# Row IQRs
q <- rowIQRs(x)
print(q)
q0 <- apply(x, MARGIN = 1, FUN = IQR)
stopifnot(all.equal(q0, q))

# Column IQRs
q <- colIQRs(x)
print(q)
q0 <- apply(x, MARGIN = 2, FUN = IQR)
stopifnot(all.equal(q0, q))
}
\seealso{
See \code{\link[stats]{IQR}}.  See \code{\link{rowSds}}().
}
