% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{colLogSumExps,dgCMatrix-method}
\alias{colLogSumExps,dgCMatrix-method}
\title{Accurately calculates the logarithm of the sum of exponentials for each row (column) of a matrix-like object}
\usage{
\S4method{colLogSumExps}{dgCMatrix}(lx, rows = NULL, cols = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{lx}{An NxK matrix-like object. Typically \code{lx} are \code{log(x)} values.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Accurately calculates the logarithm of the sum of exponentials for each row (column) of a matrix-like object
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowLogSumExps}}
/ \code{matrixStats::\link[matrixStats]{colLogSumExps}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowLogSumExps(mat)
  colLogSumExps(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowLogSumExps}()} and
\code{matrixStats::\link[matrixStats]{colLogSumExps}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item \code{\link[MatrixGenerics:rowSums2]{rowSums2()}}
}
}
