% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\docType{methods}
\name{colLogSumExps}
\alias{colLogSumExps}
\alias{colLogSumExps,dgCMatrix-method}
\alias{rowLogSumExps}
\alias{rowLogSumExps,dgCMatrix-method}
\title{Accurately computes the logarithm of the sum of exponentials across rows or
columns}
\usage{
colLogSumExps(lx, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colLogSumExps}{dgCMatrix}(lx, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)

rowLogSumExps(lx, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowLogSumExps}{dgCMatrix}(lx, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)
}
\arguments{
\item{lx}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.
Typically \code{lx} are \eqn{log(x)} values.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, any missing values are
ignored, otherwise not.}

\item{...}{Not used.}
}
\value{
A \code{\link[base]{numeric}} \code{\link[base]{vector}} of length N
(K).
}
\description{
Accurately computes the logarithm of the sum of exponentials across rows or
columns.
}
\section{Benchmarking}{


These methods are implemented in native code and have been optimized for
speed and memory.

}

\seealso{
To calculate the same on vectors, \code{\link{logSumExp}}().
}
