% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colMads,dgCMatrix-method}
\alias{colMads,dgCMatrix-method}
\alias{rowMads,dgCMatrix-method}
\title{Calculates the median absolute deviation for each row (column) of a
matrix-like object}
\usage{
\S4method{colMads}{dgCMatrix}(x, rows = NULL, cols = NULL, constant = 1.4826, na.rm = FALSE)

\S4method{rowMads}{dgCMatrix}(x, rows = NULL, cols = NULL, constant = 1.4826, na.rm = FALSE)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{constant}{A scale factor. See \code{stats::\link[stats]{mad}()} for
details.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the median absolute deviation for each row (column) of a
matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowMads}}
/ \code{matrixStats::\link[matrixStats]{colMads}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowMads(mat)
  colMads(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowMads}()} and
\code{matrixStats::\link[matrixStats]{colMads}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For mean estimates, see \code{\link[MatrixGenerics]{rowMeans2}()} and
\code{\link[base:colSums]{rowMeans}()}.
\item For non-robust standard deviation estimates, see
\code{\link[MatrixGenerics]{rowSds}()}.
}
}
