% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colMins,dgCMatrix-method}
\alias{colMins,dgCMatrix-method}
\alias{rowMins,dgCMatrix-method}
\title{Calculates the minimum for each row (column) of a matrix-like object}
\usage{
\S4method{colMins}{dgCMatrix}(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowMins}{dgCMatrix}(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the minimum for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowMins}}
/ \code{matrixStats::\link[matrixStats]{colMins}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowMins(mat)
  colMins(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowMins}()} and
\code{matrixStats::\link[matrixStats]{colMins}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For max estimates, see \code{\link[MatrixGenerics]{rowMaxs}()}.
}
}
