% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\docType{methods}
\name{colOrderStats}
\alias{colOrderStats}
\alias{colOrderStats,dgCMatrix-method}
\alias{rowOrderStats}
\alias{rowOrderStats,dgCMatrix-method}
\title{Gets an order statistic for each row (column) in a matrix}
\usage{
colOrderStats(x, rows = NULL, cols = NULL, which, ...)

\S4method{colOrderStats}{dgCMatrix}(x, rows = NULL, cols = NULL,
  which = 1, na.rm = FALSE, ...)

rowOrderStats(x, rows = NULL, cols = NULL, which, ...)

\S4method{rowOrderStats}{dgCMatrix}(x, rows = NULL, cols = NULL,
  which = 1, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{which}{An \code{\link[base]{integer}} index in [1,K] ([1,N])
indicating which order statistic to be returned.}

\item{...}{Not used.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Gets an order statistic for each row (column) in a matrix.
}
\details{
The implementation of \code{rowOrderStats()} is optimized for both speed and
memory.  To avoid coercing to \code{\link[base]{double}}s (and hence memory
allocation), there is a unique implementation for
\code{\link[base]{integer}} matrices.
}
\section{Missing values}{

 This method does \emph{not} handle missing values,
that is, the result corresponds to having \code{na.rm = FALSE} (if such an
argument would be available).

}

\seealso{
See \code{rowMeans()} in \code{\link[base]{colSums}}().
}
