% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colQuantiles,dgCMatrix-method}
\alias{colQuantiles,dgCMatrix-method}
\alias{rowQuantiles,dgCMatrix-method}
\title{Calculates quantiles for each row (column) of a matrix-like object}
\usage{
\S4method{colQuantiles}{dgCMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  probs = seq(from = 0, to = 1, by = 0.25),
  na.rm = FALSE,
  drop = TRUE
)

\S4method{rowQuantiles}{dgCMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  probs = seq(from = 0, to = 1, by = 0.25),
  na.rm = FALSE,
  drop = TRUE
)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{probs}{A numeric vector of J probabilities in [0, 1].}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{drop}{If \code{TRUE} a vector is returned if \code{J == 1}.
Note, that this is not a generic argument and not all implementation of
this function have to provide it.}
}
\value{
a \code{\link[base]{numeric}} \code{NxJ} (\code{KxJ})
\code{\link{matrix}}, where N (K) is the number of rows (columns) for
which the J values are calculated.
}
\description{
Calculates quantiles for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowQuantiles}}
/ \code{matrixStats::\link[matrixStats]{colQuantiles}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowQuantiles(mat)
  colQuantiles(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowQuantiles}()} and
\code{matrixStats::\link[matrixStats]{colQuantiles}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item \link[stats:quantile]{stats::quantile}
}
}
