% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\docType{methods}
\name{colRanges}
\alias{colRanges}
\alias{colRanges,dgCMatrix-method}
\alias{rowRanges}
\alias{rowRanges,dgCMatrix-method}
\title{Gets the range of values in each row (column) of a matrix}
\usage{
colRanges(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colRanges}{dgCMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)

rowRanges(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowRanges}{dgCMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Not used.}
}
\value{
\code{rowRanges()} (\code{colRanges()}) returns a
\code{\link[base]{numeric}} Nx2 (Kx2) \code{\link[base]{matrix}}, where N
(K) is the number of rows (columns) for which the ranges are calculated.

\code{rowMins()/rowMaxs()} (\code{colMins()/colMaxs()}) returns a
\code{\link[base]{numeric}} \code{\link[base]{vector}} of length N (K).
}
\description{
Gets the range of values in each row (column) of a matrix.
}
\seealso{
\code{\link{rowOrderStats}}() and \code{\link[base]{pmin.int}}().
}
