% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\docType{methods}
\name{colRanks}
\alias{colRanks}
\alias{colRanks,dgCMatrix-method}
\alias{rowRanks}
\alias{rowRanks,dgCMatrix-method}
\title{Gets the rank of each row (column) of a matrix}
\usage{
colRanks(x, rows = NULL, cols = NULL, ties.method = c("max",
  "average", "min"), preserve.shape = FALSE, ...)

\S4method{colRanks}{dgCMatrix}(x, rows = NULL, cols = NULL,
  ties.method = c("max", "average", "min"), preserve.shape = FALSE,
  na.handling = c("keep", "last"), ...)

rowRanks(x, rows = NULL, cols = NULL, ties.method = c("max",
  "average", "min"), preserve.shape = TRUE, ...)

\S4method{rowRanks}{dgCMatrix}(x, rows = NULL, cols = NULL,
  ties.method = c("max", "average", "min"), preserve.shape = TRUE,
  na.handling = c("keep", "last"), ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} or \code{\link[base]{integer}} NxK
\code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{ties.method}{A \code{\link[base]{character}} string specifying how
ties are treated.  For details, see below.}

\item{preserve.shape}{a boolean that specifies if the returned matrix has the same
dimensions as the input matrix. By default this is true for `rowRanks()`, but false for
`colRanks()`.}

\item{...}{Not used.}

\item{na.handling}{string specifying how `NA`s are handled. They can either be preserved with an `NA` rank
('keep') or sorted in at the end ('last'). Default is 'keep' derived from the behavior of the equivalent
`matrixStats` function.}
}
\value{
An \code{\link[base]{integer}} \code{\link[base]{matrix}} is
returned.  The \code{rowRanks()} function always returns an NxK
\code{\link[base]{matrix}}, where N (K) is the number of rows (columns)
whose ranks are calculated.

The \code{colRanks()} function returns an NxK \code{\link[base]{matrix}}, if
\code{preserveShape = TRUE}, otherwise a KxN \code{\link[base]{matrix}}.

%% The mode of the returned matrix is \code{\link[base]{integer}}, except
for %% \code{ties.method == "average"} when it is
\code{\link[base]{double}}.
}
\description{
Gets the rank of each row (column) of a matrix.
}
\details{
The row ranks of \code{x} are collected as \emph{rows} of the result matrix.

The column ranks of \code{x} are collected as \emph{rows} if
\code{preserveShape = FALSE}, otherwise as \emph{columns}.

The implementation is optimized for both speed and memory.  To avoid
coercing to \code{\link[base]{double}}s (and hence memory allocation), there
is a unique implementation for \code{\link[base]{integer}} matrices.  It is
more memory efficient to do \code{colRanks(x, preserveShape = TRUE)} than
\code{t(colRanks(x, preserveShape = FALSE))}.

Any \code{\link[base]{names}} of \code{x} are ignored and absent in the
result.
}
\section{Missing and non- values}{

 These are ranked as \code{NA}, as with
\code{na.last = "keep"} in the \code{\link[base]{rank}}() function.

}

\seealso{
\code{\link[base]{rank}}().  For developers, see also Section
'Utility functions' in 'Writing R Extensions manual', particularly the
native functions \code{R_qsort_I()} and \code{R_qsort_int_I()}.
}
