% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\docType{methods}
\name{colSums2}
\alias{colSums2}
\alias{colSums2,dgCMatrix-method}
\alias{rowSums2}
\alias{rowSums2,dgCMatrix-method}
\title{Calculates the sum for each row (column) in a matrix}
\usage{
colSums2(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colSums2}{dgCMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)

rowSums2(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowSums2}{dgCMatrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)
}
\arguments{
\item{x}{A \code{\link[base]{numeric}} or a \code{\link[base]{logical}}
NxK \code{\link[base]{matrix}}.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Not used.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the sum for each row (column) in a matrix.
}
\details{
The implementation of \code{rowSums2()} and \code{colSums2()} is
optimized for both speed and memory.
}
