% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colTabulates,dgCMatrix-method}
\alias{colTabulates,dgCMatrix-method}
\alias{rowTabulates,dgCMatrix-method}
\title{Tabulates the values in a matrix-like object by row (column)}
\usage{
\S4method{colTabulates}{dgCMatrix}(x, rows = NULL, cols = NULL, values = NULL, ...)

\S4method{rowTabulates}{dgCMatrix}(x, rows = NULL, cols = NULL, values = NULL, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{values}{the values to search for.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
a \code{\link[base]{numeric}} \code{NxJ} (\code{KxJ}) \code{\link{matrix}}, where
N (K) is the number of rows (columns) for which the J values are calculated.
}
\description{
Tabulates the values in a matrix-like object by row (column)
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowTabulates}}
/ \code{matrixStats::\link[matrixStats]{colTabulates}}.
}
\examples{
mat <- matrix(rpois(15, lambda = 3), nrow = 5, ncol = 3)
  mat[2, 1] <- NA_integer_
  mat[3, 3] <- 0L
  mat[4, 1] <- 0L
  
  print(mat)
  
  rowTabulates(mat)
  colTabulates(mat)
  
  rowTabulates(mat, values = 0)
  colTabulates(mat, values = 0)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowTabulates}()} and
\code{matrixStats::\link[matrixStats]{colTabulates}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item \code{base::\link[base]{table}()}
}
}
