% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colVars,xgCMatrix-method}
\alias{colVars,xgCMatrix-method}
\alias{rowVars,xgCMatrix-method}
\title{Calculates the variance for each row (column) of a matrix-like object}
\usage{
\S4method{colVars}{xgCMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  useNames = FALSE
)

\S4method{rowVars}{xgCMatrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  useNames = FALSE
)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link{NULL}}, no subsetting is
done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link{NA}}s
are excluded first, otherwise not.}

\item{center}{(optional) the center, defaults to the row means.}

\item{useNames}{If \code{\link{NA}}, the default behavior of the function about naming support is remained. If \code{\link{FALSE}}, no naming support is done. Else if \code{\link{TRUE}}, names attributes of result are set.}
}
\value{
Returns a \code{\link{numeric}} \code{\link{vector}} of length N (K).
}
\description{
Calculates the variance for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link{matrix}},
\code{\link{array}}, or \code{\link{numeric}} call
\code{matrixStats::rowVars}
/ \code{matrixStats::colVars}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowVars(mat)
  colVars(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowVars}()} and
\code{matrixStats::\link[matrixStats:rowVars]{colVars}()} which are used
when the input is a \code{matrix} or \code{numeric} vector.
\item For mean estimates, see \code{\link[MatrixGenerics]{rowMeans2}()} and
\code{\link[base:colSums]{rowMeans}()}.
\item For standard deviation estimates, see \code{\link[MatrixGenerics]{rowSds}()}.
\item \code{stats::\link[stats:cor]{var}()}.
}
}
