% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\docType{methods}
\name{colWeightedMads}
\alias{colWeightedMads}
\alias{colWeightedMads,dgCMatrix-method}
\alias{rowWeightedMads}
\alias{rowWeightedMads,dgCMatrix-method}
\title{Weighted Median Absolute Deviation (MAD)}
\usage{
colWeightedMads(x, w = NULL, rows = NULL, cols = NULL,
  constant = 1.4826, na.rm = FALSE, ...)

\S4method{colWeightedMads}{dgCMatrix}(x, w = NULL, rows = NULL,
  cols = NULL, constant = 1.4826, na.rm = FALSE, ties = NULL, ...)

rowWeightedMads(x, w = NULL, rows = NULL, cols = NULL,
  constant = 1.4826, na.rm = FALSE, ...)

\S4method{rowWeightedMads}{dgCMatrix}(x, w = NULL, rows = NULL,
  cols = NULL, constant = 1.4826, na.rm = FALSE, ties = NULL, ...)
}
\arguments{
\item{x}{a \code{\link[base]{numeric}} \code{\link[base]{vector}} containing
the values whose weighted MAD is to be computed.}

\item{w}{a vector of weights the same length as \code{x} giving the weights
to use for each element of \code{x}. Negative weights are treated as zero
weights. Default value is equal weight to all values.}

\item{rows}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating subset of
elements (or rows and/or columns) to operate over. If
\code{\link[base]{NULL}}, no subsetting is done.}

\item{constant}{A \code{\link[base]{numeric}} scale factor, cf.
\code{\link[stats]{mad}}.}

\item{na.rm}{a logical value indicating whether \code{\link[base]{NA}}
values in \code{x} should be stripped before the computation proceeds, or
not.  If \code{\link[base]{NA}}, no check at all for \code{\link[base]{NA}}s
is done.  Default value is \code{\link[base]{NA}} (for efficiency).}

\item{...}{Not used.}
}
\value{
Returns a \code{\link[base]{numeric}} scalar.
}
\description{
Computes a weighted MAD of a numeric vector.
}
\section{Missing values}{

 Missing values are dropped at the very beginning,
if argument \code{na.rm} is \code{\link[base:logical]{TRUE}}, otherwise not.

}

\examples{
x <- 1:10
n <- length(x)

m1 <- mad(x)
m2 <- weightedMad(x)
stopifnot(identical(m1, m2))

w <- rep(1, times = n)
m1 <- weightedMad(x, w)
stopifnot(identical(m1, m2))

# All weight on the first value
w[1] <- Inf
m <- weightedMad(x, w)
stopifnot(m  == 0)

# All weight on the first two values
w[1:2] <- Inf
m1 <- mad(x[1:2])
m2 <- weightedMad(x, w)
stopifnot(identical(m1, m2))

# All weights set to zero
w <- rep(0, times = n)
m <- weightedMad(x, w)
stopifnot(is.na(m))
}
\seealso{
For the non-weighted MAD, see \code{\link[stats]{mad}}.  Internally
\code{\link{weightedMedian}}() is used to calculate the weighted median.
}
