% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colWeightedSds,dgCMatrix-method}
\alias{colWeightedSds,dgCMatrix-method}
\alias{rowWeightedSds,dgCMatrix-method}
\title{Calculates the weighted standard deviation for each row (column) of a
matrix-like object}
\usage{
\S4method{colWeightedSds}{dgCMatrix}(x, w = NULL, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowWeightedSds}{dgCMatrix}(x, w = NULL, rows = NULL, cols = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{w}{A \code{\link{numeric}} vector of length K (N) that specifies by
how much each element is weighted.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the weighted standard deviation for each row (column) of a
matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowWeightedSds}}
/ \code{matrixStats::\link[matrixStats]{colWeightedSds}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  w <- rnorm(n = 5, mean = 3)
  rowWeightedSds(mat, w = w[1:3])
  colWeightedSds(mat, w = w)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowWeightedSds}()} and
\code{matrixStats::\link[matrixStats]{colWeightedSds}()} which are used
when the input is a \code{matrix} or \code{numeric} vector.
\item See also \link[MatrixGenerics]{rowSds} for the corresponding unweighted function.
}
}
