% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{colAnyNAs,dgCMatrix-method}
\alias{colAnyNAs,dgCMatrix-method}
\title{Check if any elements in a row (column) of a matrix-like object is missing}
\usage{
\S4method{colAnyNAs}{dgCMatrix}(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{logical}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Check if any elements in a row (column) of a matrix-like object is missing
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowAnyNAs}}
/ \code{matrixStats::\link[matrixStats]{colAnyNAs}}.
}
\examples{
  mat <- matrix(0, nrow=10, ncol=5)
  mat[sample(seq_len(5 *10), 5)] <- NA
  sp_mat <- as(mat, "dgCMatrix")
  colAnyNAs(sp_mat)

}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowAnyNAs}()} and
\code{matrixStats::\link[matrixStats]{colAnyNAs}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For checks if any element is equal to a value, see \code{\link[MatrixGenerics]{rowAnys}()}.
\item \link{is.na} and \link{any}
}
}
