% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colCumsums,dgCMatrix-method}
\alias{colCumsums,dgCMatrix-method}
\alias{rowCumsums,dgCMatrix-method}
\title{Calculates the cumulative sum for each row (column) of a matrix-like object}
\usage{
\S4method{colCumsums}{dgCMatrix}(x, rows = NULL, cols = NULL)

\S4method{rowCumsums}{dgCMatrix}(x, rows = NULL, cols = NULL)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link{NULL}}, no subsetting is
done.}

\item{cols}{A \code{\link{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link{NULL}}, no subsetting is
done.}
}
\value{
Returns a \code{\link{numeric}} \code{\link{matrix}}with the same
dimensions as \code{x}.
}
\description{
Calculates the cumulative sum for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link{matrix}},
\code{\link{array}}, or \code{\link{numeric}} call
\code{matrixStats::rowCumsums}
/ \code{matrixStats::colCumsums}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowCumsums(mat)
  colCumsums(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowCumsums}()} and
\code{matrixStats::\link[matrixStats:rowCumsums]{colCumsums}()} which are
used when the input is a \code{matrix} or \code{numeric} vector.
\item \code{base::\link{cumsum}()}.
}
}
