% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/methods_row.R
\name{colIQRs,dgCMatrix-method}
\alias{colIQRs,dgCMatrix-method}
\alias{rowIQRs,dgCMatrix-method}
\title{Calculates the interquartile range for each row (column) of a matrix-like
object}
\usage{
\S4method{colIQRs}{dgCMatrix}(x, rows = NULL, cols = NULL, na.rm = FALSE)

\S4method{rowIQRs}{dgCMatrix}(x, rows = NULL, cols = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the interquartile range for each row (column) of a matrix-like
object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowIQRs}}
/ \code{matrixStats::\link[matrixStats]{colIQRs}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowIQRs(mat)
  colIQRs(mat)
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowIQRs}()} and
\code{matrixStats::\link[matrixStats]{colIQRs}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For a non-robust analog, see \code{\link[MatrixGenerics]{rowSds}()}. For a more
robust version see \code{\link[MatrixGenerics:rowMads]{rowMads()}}
\item \code{stats::\link[stats]{IQR}()}.
}
}
